# api_server/swagger/sites.yaml
...
swagger: '2.0'
info:
    version: "1.0.0"
    title: DASPANEL Sites API
    description: |

        Site related API methods.

        The first version of the Daspanel Sites API is an exciting step forward 
        towards making an hosting control panel using Docker. We created it so 
        that you can create any site, plain or using PHP, in your local machine 
        and when ready easy deploy it on an public server exactly as you see it 
        in the development machine.

        Build something great!

    version: "1.0"
    termsOfService: http://daspanel.com/terms/
    contact:
        name: Daspanel API team
        email: admin@daspanel.com
        url: http://daspanel.com
    license:
        name: GPL-3
        url: https://opensource.org/licenses/GPL-3.0
# array of all schemes that your API supports
schemes:
    - http
    - https
consumes:
    - application/json
produces:
    - application/json
security:
    - internalApiKey: []
securityDefinitions:
    internalApiKey:
        type: apiKey
        in: header
        name: X-Api-Key

#tags:
#    - name: Databases
#      description: |
#          Relationships are expressed using the following terms:
#
#          **outgoing_status**: Your relationship to the user. Can be "follows",
#            "requested", "none".
#          **incoming_status**: A user's relationship to you. Can be "followed_by",
#            "requested_by", "blocked_by_you", "none".
#    - name: Actions
#      description: |
#            At this time, uploading via the API is not possible. We made a conscious
#            choice not to add this for the following reasons:
#
#            * Instagram is about your life on the go – we hope to encourage photos
#            from within the app.
#            * We want to fight spam & low quality photos. Once we allow uploading
#            from other sources, it's harder to control what comes into the Instagram
#            ecosystem. All this being said, we're working on ways to ensure users
#            have a consistent and high-quality experience on our platform.
paths:
...
    /service/{servertype}/genconfig/{cuid}:
        get:
            tags: [Servers]
            operationId: modules.sites.serviceapi.httpserver_gencfg
            summary: Generate config files for site server app
            description: |
                This endpoint is for generate service app config of a site.
            parameters:
                - $ref: '#/parameters/cuid'
                - $ref: '#/parameters/servertype'
            responses:
                200:
                    description: Server config.
                    schema:
                        $ref: '#/definitions/RedirectsItems'
                400:
                    $ref: '#/responses/400BadRequest'
                401:
                    $ref: '#/responses/401Unauthorized'
                403:
                    $ref: '#/responses/403Forbidden'
                404:
                    $ref: '#/responses/404NotFound'
                429:
                    $ref: '#/responses/429TooManyRequests'
                500:
                    $ref: '#/responses/500InternalServerError'
                503:
                    $ref: '#/responses/503ServiceUnavailable'
    /service/{servertype}/activate/{cuid}:
        get:
            tags: [Servers]
            operationId: modules.sites.serviceapi.httpserver_activate
            summary: Activate site in HTTP server app
            description: |
                This activate site in HTTP server app.
            parameters:
                - $ref: '#/parameters/cuid'
                - $ref: '#/parameters/servertype'
            responses:
                200:
                    $ref: '#/responses/Success'
                400:
                    $ref: '#/responses/400BadRequest'
                401:
                    $ref: '#/responses/401Unauthorized'
                403:
                    $ref: '#/responses/403Forbidden'
                404:
                    $ref: '#/responses/404NotFound'
                429:
                    $ref: '#/responses/429TooManyRequests'
                500:
                    $ref: '#/responses/500InternalServerError'
                503:
                    $ref: '#/responses/503ServiceUnavailable'
    /service/{servertype}/deactivate/{cuid}:
        get:
            tags: [Servers]
            operationId: modules.sites.serviceapi.httpserver_deactivate
            summary: Deactivate site in HTTP server app
            description: |
                This deactivate site in HTTP server app.
            parameters:
                - $ref: '#/parameters/cuid'
                - $ref: '#/parameters/servertype'
            responses:
                200:
                    $ref: '#/responses/Success'
                400:
                    $ref: '#/responses/400BadRequest'
                401:
                    $ref: '#/responses/401Unauthorized'
                403:
                    $ref: '#/responses/403Forbidden'
                404:
                    $ref: '#/responses/404NotFound'
                429:
                    $ref: '#/responses/429TooManyRequests'
                500:
                    $ref: '#/responses/500InternalServerError'
                503:
                    $ref: '#/responses/503ServiceUnavailable'
    /service/{servertype}/reload:
        get:
            tags: [Servers]
            operationId: modules.sites.serviceapi.httpserver_reload
            summary: Reload HTTP server app configuration
            description: |
                This reload HTTP server app configuration.
            parameters:
                - $ref: '#/parameters/servertype'
            responses:
                200:
                    $ref: '#/responses/Success'
                400:
                    $ref: '#/responses/400BadRequest'
                401:
                    $ref: '#/responses/401Unauthorized'
                403:
                    $ref: '#/responses/403Forbidden'
                404:
                    $ref: '#/responses/404NotFound'
                429:
                    $ref: '#/responses/429TooManyRequests'
                500:
                    $ref: '#/responses/500InternalServerError'
                503:
                    $ref: '#/responses/503ServiceUnavailable'
...
parameters:
    servertype:
        name: servertype
        description: "Daspanel service server type, like: nginx"
        in: path
        type: string
        required: true
        minLength: 1
        maxLength: 32
...
    cuid:
        name: cuid
        description: Item Universal unique identifier
        in: path
        type: string
        required: true
        pattern: "^[a-zA-Z0-9-]+$"
...
definitions:
...
    RedirectsItems:
        type: array
        items:
            $ref: '#/definitions/SiteRedirect'
    SiteRedirect:
        type: object
        properties:
            hosturl:
                type: string
                description: Redirect host
                example: "www"
                minLength: 1
                maxLength: 255
            domain:
                type: string
                description: Domain of the redirect
                example: "mydomain.com"
                minLength: 1
                maxLength: 255
            ssl:
                type: boolean
                description: Redirect had SSL cert
            sslcert:
                type: string
                description: Directory of SSL cert
                example: "data/cit3gxhqh00007aruv60es3XX/certs/siteuuid-sdlkslkj/certuuid-09sdhsdojhsd"
                minLength: 1
                maxLength: 255
            _cuid:
                type: string
                description: Universal unique id of the redirect
                example: "cir22a51g0005mw45ay17gkeh"
                minLength: 25
                maxLength: 25
                readOnly: true
    Message:
        type: object
        required:
            - status
            - type
            - title
            - detail
        properties:
            status:
                description: HTTP status code
                type: integer
                format: int32
                readOnly: true
            type:
                description: Url where the probem is documented
                type: string
                format: url
                readOnly: true
            title:
                description: Title of the problem
                type: string
                readOnly: true
            detail:
                description: Details about the problem
                type: string
                readOnly: true
...
responses:
    Success:
        description: Success
...
    400BadRequest:
        description: The request was invalid or cannot be otherwise served.
        schema:
            $ref: '#/definitions/Message'
    401Unauthorized:
        description: Authentication credentials were missing or incorrect.
        schema:
            $ref: '#/definitions/Message'
    403Forbidden:
        description: The request is understood, but it has been refused or access is not allowed.
        schema:
            $ref: '#/definitions/Message'
    404NotFound:
        description: The URI requested is invalid or the resource requested does not exists. 
        schema:
            $ref: '#/definitions/Message'
    429TooManyRequests:
        description: The request cannot be served due to the application’s rate limit having been exhausted for the resource. 
        schema:
            $ref: '#/definitions/Message'
    500InternalServerError:
        description: Something is broken.
        schema:
            $ref: '#/definitions/Message'
    503ServiceUnavailable:
        description: The server is up, but overloaded with requests. Try again later. 
        schema:
            $ref: '#/definitions/Message'